Move matcher to sub-packages
