--- app/src/main/java/com/realityexpander/tasky/auth_feature/domain/validation/validateEmail/IEmailMatcher.kt
@@ -1,5 +1,0 @@
-package com.realityexpander.tasky.auth_feature.domain.validation.validateEmail
-
-interface IEmailMatcher {
-    fun matches(email: String): Boolean
-}
