+++ app/src/main/java/com/realityexpander/tasky/auth_feature/domain/validation/validateEmail/emailMatcher/IEmailMatcher.kt
@@ -1,0 +1,5 @@
+package com.realityexpander.tasky.auth_feature.domain.validation.validateEmail.emailMatcher
+
+interface IEmailMatcher {
+    fun matches(email: String): Boolean
+}
