+++ app/src/main/java/com/realityexpander/tasky/auth_feature/domain/validation/validateEmail/validateEmailImpls/ValidateEmailRegexImpl.kt
@@ -1,8 +1,8 @@
 package com.realityexpander.tasky.auth_feature.domain.validation.validateEmail.validateEmailImpls
 
-import com.realityexpander.tasky.auth_feature.domain.validation.validateEmail.IEmailMatcher
 import com.realityexpander.tasky.auth_feature.domain.validation.validateEmail.IValidateEmail
-import com.realityexpander.tasky.auth_feature.domain.validation.validateEmail.emailMatcherImpls.EmailMatcherRegexImpl
+import com.realityexpander.tasky.auth_feature.domain.validation.validateEmail.emailMatcher.IEmailMatcher
+import com.realityexpander.tasky.auth_feature.domain.validation.validateEmail.emailMatcher.emailMatcherImpls.EmailMatcherRegexImpl
 
 class ValidateEmailRegexImpl (
     private val emailMatcher: IEmailMatcher = EmailMatcherRegexImpl()
